move ElementDispatcher in Gaphor.UML
This functionality is very model related (used by gaphor.diagram code) and therefore should deserve a place in gaphor.UML.
