+++ gaphor/diagram/general/generalpropertypages.py
@@ -2,7 +2,7 @@ from gi.repository import Gtk
 
 from gaphor import UML
 from gaphor.core import _, transactional
-from gaphor.services.elementdispatcher import EventWatcher
+from gaphor.UML.elementdispatcher import EventWatcher
 from gaphor.diagram.propertypages import PropertyPages, PropertyPageBase
