+++ gaphor/diagram/profiles/metaclasspropertypage.py
@@ -6,7 +6,7 @@ from gi.repository import Gtk
 
 from gaphor import UML
 from gaphor.core import _
-from diagram.services.elementdispatcher import EventWatcher
+from gaphor.UML.elementdispatcher import EventWatcher
 from gaphor.diagram.propertypages import create_hbox_label
 from gaphor.diagram.propertypages import PropertyPages, NamedElementPropertyPage
 from gaphor.diagram.classes import ClassItem
