+++ gaphor/diagram/propertypages.py
@@ -33,7 +33,7 @@ from gi.repository import Gtk
 
 from gaphor import UML
 from gaphor.core import _, transactional
-from gaphor.services.elementdispatcher import EventWatcher
+from gaphor.UML.elementdispatcher import EventWatcher
 
 
 class _PropertyPages:
